Overlay - Added New APIs
Added the API commitUpdates() to overlay_control_device_t. This API must be called after a setPosition or setParameter call, for the changes to take
effect. In this way, we avoid multiple stream on/off calls and the transition is more smooth.

Added the API resizeInput() to overlay_data_device_t. There are cases where the Camera HAL will be receiving preview frames from the image pipe that
are actually larger than the client requested size. At the moment, the overlay architecture does not allow the data side of overlay to dictate a
preview input frame size. As an interim solution, we have added a new overlay data side API that would allow the Camera HAL to resize the input frame
size.

Added the API setCrop() to overlay_data_device_t. This API can be used to set the crop parameters.

===============================================

Currently, the display driver allocates cacheable buffers always.
Cacheable buffers are needed mainly for video record use case
with VSTAB(Video Stablilization). VSTAB runs on ARM and hence
works faster with cached buffers. Because of these cacheable
buffers, display driver has to flush the cache before rendering.
But for video playback use case we don't need cached buffers,
so this flush operation in driver is an overhead. We can improve
power and performance numbers by avoiding the flush operation
in driver. Therefore, we want to make this(cacheable buffers)
a configurable option. So, we would like to add the following
DATA API to the overlay module:

int overlay_setAttributes( struct overlay_data_device_t *dev,
int param, int value)

This interface will have to be available as data side API
because CameraHAL or video MIO node would know what the use
case is and can set required buffer attributes. The attributes
which can be set would be cacheable or non cacheable buffers,
physically contiguous buffers, etc.

===============================================
This patch is equivalent to Google Patches 9888 and 9899.
===============================================

Signed-off-by: Anu Sundararajan <sanuradha@ti.com>