libhardware: OMAP extension to extend hwc api
Support an extension to the HWC callback mechanism, so
that the api can be freely extended in the future.

This change is intended to keep ABI compatibility for
the HWC as this api will only be used if HWC_EXTENDED_API
is set in the flags.

This isn't bullet proof ABI protection as it assumes
that both SurfaceFlinger and HWC implementation will
support the same extensions.

Change-Id: I257c4da4b3a474119521cb03666b6a18ba58fc66
Signed-off-by: Tony Lofthouse <a0741364@ti.com>