hwc: Extended API for querying display layer stack
Upon connection of external display SurfaceFlinger assigns it the
same layer stack as for the primary display to enable cloning. The
primary layer stack is then rendered for each display separately,
which doubles GPU load. In most cases cloning can be done more
efficiently in HWC, but in order to do it HWC should know which
displays are showing cloned composition. By querying layer stack
for each display HWC can decide if it should do the cloning or not.

Change-Id: Ib5166d6b54e2a132f7c031ac98dba992c62ffb5f
Signed-off-by: Mykola Ostrovskyy <mykola@ti.com>